Move tool host_cross_arch tests into different shards
