+++ dev/bots/test.dart
@@ -230,14 +230,8 @@ Future<void> _runWebToolTests() async {
   );
 }
 
-Future<void> _runToolHostCrossArchTests() {
-  return runDartTest(
-    _toolsPath,
-    // These are integration tests
-    forceSingleCore: true,
-    testPaths: <String>[path.join('test', 'host_cross_arch.shard')],
-  );
-}
+// TODO(jmagman): https://github.com/flutter/flutter/issues/189302 remove when it gets to stable.
+Future<void> _runToolHostCrossArchTests() async {}
 
 Future<void> _runIntegrationToolTests() async {
   final List<String> allTests = Directory(path.join(_toolsPath, 'test', 'integration.shard'))
